Refactor login page
